chore(pairing): retire vestigial node token plumbing and flatten cleanup claims
