+++ src/infra/pairing-files.ts
@@ -37,46 +37,3 @@ export function pruneExpiredPending<T extends { ts: number; refreshedAtMs?: numb
     }
   }
 }
-
-/** Result shape for creating or refreshing a pending pairing request. */
-export type PendingPairingRequestResult<TPending> = {
-  status: "pending";
-  request: TPending;
-  created: boolean;
-};
-
-/** Refresh one compatible pending request or replace a superseded request set atomically. */
-export async function reconcilePendingPairingRequests<
-  TPending extends { requestId: string },
-  TIncoming,
->(params: {
-  pendingById: Record<string, TPending>;
-  existing: readonly TPending[];
-  incoming: TIncoming;
-  canRefreshSingle: (existing: TPending, incoming: TIncoming) => boolean;
-  refreshSingle: (existing: TPending, incoming: TIncoming) => TPending;
-  buildReplacement: (params: { existing: readonly TPending[]; incoming: TIncoming }) => TPending;
-  persist: () => Promise<void>;
-}): Promise<PendingPairingRequestResult<TPending>> {
-  if (
-    params.existing.length === 1 &&
-    params.canRefreshSingle(params.existing[0], params.incoming)
-  ) {
-    const refreshed = params.refreshSingle(params.existing[0], params.incoming);
-    params.pendingById[refreshed.requestId] = refreshed;
-    await params.persist();
-    return { status: "pending", request: refreshed, created: false };
-  }
-
-  for (const existing of params.existing) {
-    delete params.pendingById[existing.requestId];
-  }
-
-  const request = params.buildReplacement({
-    existing: params.existing,
-    incoming: params.incoming,
-  });
-  params.pendingById[request.requestId] = request;
-  await params.persist();
-  return { status: "pending", request, created: true };
-}
